#7, move enum to model, addition of comment on manager, removal of empty enum ReplicaAddress
