+++ Replica-Manager/Manager/ManagerMessageHandler.java
@@ -1,6 +1,8 @@
 package Manager;
 
 import UDP.Message;
+import model.ManagersMap;
+import model.RegisteredReplica;
 
 /**
  * Handles message , parse if needed and redirect to the right Replica Manager
@@ -9,30 +11,38 @@ import UDP.Message;
  */
 public class ManagerMessageHandler {
 	
+	public ManagerMessageHandler() {
+		
+	}
 	public String HandleRequest(Message msg) {
 		
+		//Find out which Replica is targeted
 		RegisteredReplica replicaFound = parseFindReplicaName(msg.getData());
 		if(replicaFound == null) {
 			//TODO: log this, error
-			System.out.println("Was not able to find the replica name";);
+			System.out.println("Was not able to find the replica name");
 			return null;
 		}
+		
+		// Grab the manager current instance
+		ManagersMap currentManagers = ManagersMap.getManagers();
+		Manager currentManager = currentManagers.get(replicaFound);
+		
+		// Do action based on msg...
 		switch(msg.getOpCode()) {
 		case NO_RESP_NOTIFICATION:
 			System.out.println("No Resp Notif");
-			
-			break;
-		case ACK_NO_RESP_NOTIFICATION:
-			System.out.println("Ack Resp");
+			//TODO: Add logging here
+			currentManager.registerCrashFailure(msg.getSeqNum());
 			break;
 		case FAULY_RESP_NOTIFICATION:
 			System.out.println("FAULY RESP Notification");
-			break;
-		case ACK_FAULY_RESP_NOTIFICATION:
-			System.out.println("Acknowledge Fauly Resp. Notif");
+			currentManager.registerNonByzFailure(msg.getSeqNum());
+			//TODO: Add logging here
 			break;
 		default:
 			System.out.println("Operation not recognized");
+			//TODO: Add logging here
 			break;
 		}
 		return null;
