--- Replica-Manager/model/ReplicaAdress.java
@@ -1,5 +1,0 @@
-package model;
-
-public enum ReplicaAdress {
-
-}
